cursesman/server_requirements.txt
tornado==6.1
python-socketio==5.0.4
